requirements.txt
google-genai>=0.2.0
pyaudio>=0.2.14
python-dotenv>=1.0.1
